Consolidate riskfacs
